add img_vectorize
