+++ nodeList.js
@@ -4,9 +4,9 @@ import async_adder from "./nodes/async_adder.js";
 import img_distance from "./nodes/img_distance.js";
 import img_edge_detect from "./nodes/img_edge_detect.js";
 import img_offset from "./nodes/img_offset.js";
-import img_read_png from "./nodes/img_read_png.js";
 import img_sample from "./nodes/img_sample.js";
 import img_threshold from "./nodes/img_threshold.js";
+import img_vectorize from "./nodes/img_vectorize.js";
 import img_viewer from "./nodes/img_viewer.js";
 import number from "./nodes/number.js";
 import worker_example from "./nodes/worker-example.js";
@@ -18,6 +18,7 @@ export default {
   img_offset,
   img_sample,
   img_threshold,
+  img_vectorize,
   img_viewer,
   number,
   worker_example
